modify stream now supported in api v2
